refactor(chains): extract summarization chain builder
